fix(openfold3): tighten qualification contracts
Own atom matching in the model family, require complete aligned-structure coverage for mixed FP16, and make E2E precision expectations manifest-driven.
